runtime: Remove `module` mod
