+++ crates/rt/src/workflow/persistence.rs
@@ -11,8 +11,8 @@ use crate::{
         Channels, ChildWorkflow, ConsumeError, PersistError, PersistedWorkflowData, ReceiverState,
         SenderState, TaskState, TimerState, Wakers,
     },
-    engine::{CreateWorkflow, PersistWorkflow, RunWorkflow},
-    module::{Services, WorkflowSpawner},
+    engine::{CreateWorkflow, PersistWorkflow, RunWorkflow, WorkflowSpawner},
+    manager::Services,
     receipt::WakeUpCause,
     utils::Message,
     workflow::{ChannelIds, Workflow},
